server/requirements.txt
flask-cors==4.0.1
python-dotenv==1.0.1
stripe==9.6.0
